test-requirements.txt
# Python packages required for testing this collection
pytest-ansible
pytest-xdist
molecule
# Security testing
bandit
safety
# Code quality
black
flake8
isort
